enablement: remove recipe projection import cycle
